+++ src/routes/Game/routes/Board/index.js
@@ -1,0 +1,24 @@
+import s from './style.module.css';
+
+const BoardPage = () => {
+    return (
+        <div className={s.root}>
+						<div className={s.playerOne}>
+
+						</div>
+            <div className={s.board}>
+                <div className={s.boardPlate}>1</div>
+                <div className={s.boardPlate}>2</div>
+                <div className={s.boardPlate}>3</div>
+                <div className={s.boardPlate}>4</div>
+                <div className={s.boardPlate}>5</div>
+                <div className={s.boardPlate}>6</div>
+                <div className={s.boardPlate}>7</div>
+                <div className={s.boardPlate}>8</div>
+                <div className={s.boardPlate}>9</div>
+            </div>
+        </div>
+    );
+};
+
+export default BoardPage;
