+++ src/routes/Game/routes/Finish/index.js
@@ -1,0 +1,7 @@
+const FinishPage = () => {
+    return (
+        <h1>This is Finish Page</h1>
+    );
+};
+
+export default FinishPage;
